+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -25,6 +25,11 @@
   [1.42.0](https://github.com/open-telemetry/semantic-conventions/blob/v1.42.0/docs/rpc/README.md).
   ([#4508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4508))
 
+* Fixed gRPC attributes being missing from the exported span when a sibling
+  activity is created because a non-default propagator (e.g. a custom or
+  composite propagator) is configured.
+  ([#4508](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4508))
+
 ## 1.15.2
 
 Released 2026-Apr-21